chore: finishes css pre-compilation
