+++ package.json
@@ -9,7 +9,6 @@
   "scripts": {
     "build": "pnpm run build:core",
     "build:all": "turbo build",
-    "build:webpack2": "webpack --config webpack.config.mjs",
     "build:core": "turbo build --filter \"!@payloadcms/plugin-*\"",
     "build:plugins": "turbo build --filter \"@payloadcms/plugin-*\" --filter \"!@payloadcms/plugin-search\" --filter \"!@payloadcms/plugin-redirects\" --filter \"!@payloadcms/plugin-nested-docs\"",
     "build:app": "next build",
@@ -72,7 +71,6 @@
     "@octokit/core": "^5.1.0",
     "@payloadcms/eslint-config": "workspace:*",
     "@playwright/test": "1.42.1",
-    "@swc-node/loader": "^1.5.0",
     "@swc/cli": "^0.1.62",
     "@swc/jest": "0.2.36",
     "@testing-library/jest-dom": "6.4.2",
@@ -102,8 +100,6 @@
     "conventional-changelog-writer": "^7.0.1",
     "copyfiles": "2.4.1",
     "cross-env": "7.0.3",
-    "css-loader": "^6.10.0",
-    "css-minimizer-webpack-plugin": "^6.0.0",
     "dotenv": "8.6.0",
     "drizzle-orm": "0.29.4",
     "execa": "5.1.1",
@@ -119,7 +115,6 @@
     "jwt-decode": "3.1.2",
     "lexical": "0.13.1",
     "lint-staged": "^14.0.1",
-    "mini-css-extract-plugin": "1.6.2",
     "minimist": "1.2.8",
     "mongodb-memory-server": "^9",
     "next": "14.1.2",
@@ -130,8 +125,6 @@
     "pino-pretty": "10.2.0",
     "playwright": "file:playwright-1.43.0-next.tgz",
     "playwright-core": "file:playwright-core-1.43.0-next.tgz",
-    "postcss-loader": "^8.1.1",
-    "postcss-preset-env": "^9.5.0",
     "prettier": "^3.0.3",
     "prompts": "2.4.2",
     "qs": "6.11.2",
@@ -139,24 +132,18 @@
     "react-dom": "^18.2.0",
     "read-stream": "^2.1.1",
     "rimraf": "3.0.2",
-    "sass-loader": "^14.1.1",
     "semver": "^7.5.4",
     "sharp": "0.32.6",
     "shelljs": "0.8.5",
     "simple-git": "^3.20.0",
     "slash": "3.0.0",
     "slate": "0.91.4",
-    "swc-loader": "^0.2.6",
     "tempfile": "^3.0.0",
-    "terser-webpack-plugin": "^5.3.10",
-    "ts-loader": "^9.5.1",
     "ts-node": "10.9.1",
     "tsx": "^4.7.1",
     "turbo": "^1.12.5",
     "typescript": "5.2.2",
     "uuid": "^9.0.1",
-    "webpack": "^5.78.0",
-    "webpack-cli": "^5.1.4",
     "yocto-queue": "^1.0.0"
   },
   "peerDependencies": {
