+++ packages/plugin-cloud/.swcrc
@@ -1,6 +1,6 @@
 {
   "$schema": "https://json.schemastore.org/swcrc",
-  "sourceMaps": "inline",
+  "sourceMaps": true,
   "exclude": ["/**/*.spec.ts", "/**/mocks"],
   "jsc": {
     "target": "esnext",
