+++ packages/plugin-sentry/.swcrc
@@ -1,6 +1,6 @@
 {
   "$schema": "https://json.schemastore.org/swcrc",
-  "sourceMaps": "inline",
+  "sourceMaps": true,
   "exclude": ["/**/*.spec.ts"],
   "jsc": {
     "target": "esnext",
