+++ packages/ui/src/scss/app.scss
@@ -3,6 +3,10 @@
 @import './toastify.scss';
 @import './colors.scss';
 
+.gordon-freeman {
+  color: red;
+}
+
 :root {
   --base-px: 25;
   --base-body-size: 13;
